+++ Source/SysPSP/Utility/FastMemcpyPSP.cpp
@@ -309,7 +309,7 @@ void memcpy_vfpu_byteswap( void* dst, const void* src, size_t size )
 
 				u32 srcTmp {*src32++};
 				u32 dstTmp {0};
-				u32 size32 {size >> 2}; // Size in dwords (to avoid a sltiu in loop)
+				u32 size32 {static_cast<u32>(size >> 2)}; // Size in dwords (to avoid a sltiu in loop)
 
 				size &= 0x3; // Update remaining bytes if any..
